requirements.txt
vosk==0.3.45
pyaudio==0.2.14
pyyaml==6.0.1
numpy==1.24.3
nltk==3.8.1
scikit-learn==1.3.0
sentence-transformers==2.2.2
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
